Change shape to usize
The dimensions are bound by the length of the data vector, hence they are bound by usize::MAX.
Imagine this library being used on a 16-bit architecture and a 128-bit architecture. Would it make sense that the shape is always u64, even though less (or more) elements can be allocated in the data vector? In my opinion, no.

+++ benches/benchmarks.rs
@@ -8,7 +8,7 @@ use rand::{
 };
 use tetra::{contract, Tensor};
 
-fn random_tensor(shape: &[u64]) -> Tensor {
+fn random_tensor(shape: &[usize]) -> Tensor {
     let mut rng = StdRng::seed_from_u64(0);
     let range = Uniform::new(-10.0, 10.0);
     let number_elements = Tensor::total_items(shape);
